+++ source/pony.web/src/main/java/net/dorokhov/pony/web/server/service/AlbumServiceFacade.java
@@ -7,13 +7,13 @@ import java.util.ArrayList;
 
 public interface AlbumServiceFacade {
 
-	public Long getCountByArtist(Integer aArtistId);
+	public Long getCountByArtist(Long aArtistId);
 
-	public ArrayList<AlbumSongsDto> getByArtist(Integer aArtistId);
+	public ArrayList<AlbumSongsDto> getByArtist(Long aArtistId);
 	public ArrayList<AlbumSongsDto> getByArtistIdOrName(String aIdOrName);
 
 	public ArrayList<AlbumDto> search(String aQuery);
 
-	public AlbumSongsDto getById(Integer aId);
+	public AlbumSongsDto getById(Long aId);
 
 }
